add the autonumber icon and hiding dlt and duplicate

+++ src/table/Header.js
@@ -12,6 +12,8 @@ import MultiIcon from "./img/Multi";
 import HashIcon from "./img/Hash";
 import PlusIcon from "./img/Plus";
 import { shortId } from "./utils";
+// import FormatListNumberedRtlIcon from '@mui/icons-material/FormatListNumberedRtl';
+import PlusOneIcon from '@mui/icons-material/PlusOne';
 import CheckIcon from '@mui/icons-material/Check';
 import DateRangeIcon from '@mui/icons-material/DateRange';
 import AttachFileIcon from '@mui/icons-material/AttachFile';
@@ -352,7 +354,7 @@ export default function Header({
    
   ];
 
-  if (dataType !== "createdat" && dataType !== "createdby" && dataType !== "id") {
+  if (dataType !== "createdat" && dataType !== "createdby" && dataType !== "rowid" && dataType !== "autonumber") {
     buttons.push({
       onClick: () => {
         setExpanded(false);
@@ -363,7 +365,7 @@ export default function Header({
     });
   }
   
-  if (dataType !== "createdat" && dataType !== "createdby" && dataType !== "rowid") {
+  if (dataType !== "createdat" && dataType !== "createdby" && dataType !== "rowid" && dataType !== "autonumber") {
     buttons.push({
       onClick: () => {
         dispatch(deleteColumns({
@@ -436,9 +438,11 @@ export default function Header({
       case "multipleselect":
         propertyIcon = <QueueOutlinedIcon fontSize="2px" />;
         break;
-      
+      case "autonumber":
+        propertyIcon = <PlusOneIcon fontSize="2px" />;
+        break;
 
-    default:
+     default:
       propertyIcon = <MultiIcon />;
       break;
   }
